Cancel itinerary if the blueprint is triggered too frequently while in the activation zone (#37)
If the blueprint is triggered too frequently in the activation zone, it will stop and send a notification both to the driver and on HA's dashboard.
If this happens 3 times in a row, the blueprint will disable itself until the user decides to manually enable it back.

This prevents the gate from opening and closing in a loop if the driving sensor is not working properly (e.g. a Bluetooth device too far away).